+++ packages/waku/src/lib/config.ts
@@ -1,5 +1,5 @@
 import type { Config, ResolvedConfig } from '../config.js';
-import { normalizePath } from './middleware/rsc/utils.js';
+import { normalizePath } from './utils/path.js';
 
 const splitHTML = (htmlStr: string): readonly [string, string, string] => {
   const P1 = [
