+++ packages/waku/src/lib/middleware/rsc/utils.ts
@@ -1,10 +1,5 @@
 // This file should not include Node specific code.
 
-// Refs: https://github.com/rollup/plugins/blob/d49bbe8dc5ec41157de5787c72c858f73be107ff/packages/pluginutils/src/normalizePath.ts
-export function normalizePath(filePath: string) {
-  return filePath.split('\\').join('/');
-}
-
 export const encodeInput = (input: string) => {
   if (input === '') {
     return '_';
@@ -60,26 +55,6 @@ export const deepFreeze = (x: unknown): void => {
       deepFreeze(value);
     }
   }
-};
-
-export const endStream = async (stream: WritableStream, message?: string) => {
-  const writer = stream.getWriter();
-  await writer.ready;
-  if (message) {
-    await writer.write(new TextEncoder().encode(message));
-  }
-  await writer.close();
-};
-
-export const concatUint8Arrays = (arrs: Uint8Array[]): Uint8Array => {
-  const len = arrs.reduce((acc, arr) => acc + arr.length, 0);
-  const array = new Uint8Array(len);
-  let offset = 0;
-  for (const arr of arrs) {
-    array.set(arr, offset);
-    offset += arr.length;
-  }
-  return array;
 };
 
 // TODO is this correct? better to use a library?
